Implement Home screen (replaces placeholder screen)
The Home screen was just a placeholder screen and the new UI replaces
it. It includes a simple image & footer text, and includes a footer
Learn More button that opens a web screen of codeandrobots.com so
the user can find out more about the Code&Robots project.

List of changes:
 - Add Hello.png (1x, 2x, 3x) into App/Images
 - Add RobotoMedium.ttf and RobotoBold.ttf fonts, and copied them to
   Android and XCode projects
 - Update status bar background colour and style
 - Update navigation style
 - Implement Hello screen UI
 - Add onLearnMorePress in Home container to show codeandrobots.com
 - Set Home screen title
 - Fix console.log in ScreenTrackingMiddleware
 - Refactor Themes/Colors, Themes/Metrics and Themes/Fonts
 - Add NSAppTransportSecurity exception for codeandrobots.com
 - Update Home screen test
 - Up app version for Android